Create library with pickle functionality, called in main
