+++ Cargo.toml
@@ -23,6 +23,10 @@ colored = "2.0.0"
 pulldown-cmark = "0.8.0"
 rayon = "1.3"
 
+[lib]
+name = "morty"
+path = "src/lib.rs"
+
 [patch.crates-io]
 # sv-parser = {path = "sv-parser/sv-parser"}
 # sv-parser = { git = "https://github.com/zarubaf/sv-parser" }
